Add testing to CI
It includes testing on Linux, Windows, MacOS along with Python
versions 3.8, 3.9, 3.10, 3.11, 3.12. The tests install from the
built distribution to make sure we are testing against what customers
would be installing from PyPI. As of right now, there is just a single
test to make sure the installed package is importable.