Add io.BlobIO interface
It is a file-like object for reading and writing to Azure Blob
Storage. This commit mainly establishes the interface and does not
fully implement read() or write(). For read(), it relies on a
stubbed downloader interface for the basis for tests. In subsequent
commits, the read() and write() will be implemented.